refactor: split index.ts into src/cli/commands/ modules
Split 3,988-line src/index.ts into focused modules:
- src/cli/types.ts: GlobalOptions interface
- src/cli/utils.ts: shared proxy/container/port utilities
- src/cli/commands/*.ts: 29 individual command handlers
- src/cli/index.ts: main() dispatch loop
- src/index.ts: 1-line barrel shim (export * from cli/index.js)

Also adds OpenSpec proposal/design/specs/tasks for Phase 2.

Zero behavior changes. Tests: 1518 passing (10 pre-existing watcher failures).
