+++ openspec/changes/nano-brain-phase2-cli-split/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-04
