+++ src/cli/types.ts
@@ -1,0 +1,5 @@
+export interface GlobalOptions {
+  dbPath: string;
+  configPath: string;
+  remaining: string[];
+}
